+++ odbc-api/src/handles/column_description.rs
@@ -1,4 +1,4 @@
-use super::{data_type::DataType, sql_char::{SqlChar, DecodingError, vec_to_utf8}};
+use super::{data_type::DataType, sql_char::{SqlChar, DecodingError, slice_to_utf8}};
 
 /// Indication of whether a column is nullable or not.
 #[derive(Clone, Copy, Hash, Debug, Eq, PartialEq)]
@@ -45,7 +45,7 @@ impl ColumnDescription {
     /// Converts the internal UTF16 representation of the column name into UTF8 and returns the
     /// result as a `String`.
     pub fn name_to_string(&self) -> Result<String, DecodingError> {
-        vec_to_utf8(&self.name)
+        slice_to_utf8(&self.name)
     }
 
     /// `true` if the column is `Nullable` or it is not know whether the column is nullable. `false`